+++ lib/bh/helpers/modal_helper.rb
@@ -39,6 +39,7 @@ module Bh
       modal = Bh::Modal.new self, *args, &block
       modal.extract! :button, :size, :body, :title, :id
 
+      modal.extract_from :button, [:context, :size, :layout, :caption]
       modal.append_class_to! :button, :btn
       modal.append_class_to! :button, modal.button_context_class
       modal.append_class_to! :button, modal.button_size_class